Split WagtailReactModal into separate component
